backend/requirements.txt
asgiref==3.3.4
azure-ai-formrecognizer==3.1.1
azure-common==1.1.27
azure-core==1.15.0
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
djangorestframework==3.12.4
idna==2.10
install==1.3.4
isodate==0.6.0
msrest==0.6.21
oauthlib==3.1.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.5
tensorflow==1.13.2
